# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Ticketing API Documentation",
        "description": "Creation of an API in PHP Laravel for event ticketing",
        "contact": {
            "email": "ronsard2014@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get all events",
                "description": "Retrieve the list of upcoming events",
                "operationId": "GetEvent",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    },
                    "500": {
                        "description": "Unable to fetch events"
                    }
                }
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Create a new event",
                "description": "Create a new event with the provided details.",
                "operationId": "StoreEvent",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "event_category",
                                    "event_title",
                                    "event_date",
                                    "event_city",
                                    "event_address",
                                    "event_status"
                                ],
                                "properties": {
                                    "event_category": {
                                        "type": "string",
                                        "example": "Music"
                                    },
                                    "event_title": {
                                        "type": "string",
                                        "example": "Concert de Rock"
                                    },
                                    "event_description": {
                                        "type": "string",
                                        "example": "Un concert de rock avec les plus grandes stars."
                                    },
                                    "event_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-09-15T20:00:00Z"
                                    },
                                    "event_image": {
                                        "type": "string",
                                        "example": "https://example.com/images/concert.jpg"
                                    },
                                    "event_city": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "event_address": {
                                        "type": "string",
                                        "example": "Stade de France, 93216 Saint-Denis"
                                    },
                                    "event_status": {
                                        "type": "string",
                                        "example": "upcoming"
                                    },
                                    "event_created_on": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-08-14T12:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket type created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/v1/events/{event}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get event by ID",
                "description": "Retrieve details of a specific event by its ID.",
                "operationId": "ShowEvent",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID of the event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Update an existing event",
                "description": "Update details of an existing event by its ID.",
                "operationId": "UpdateEvent",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID of the event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "event_category",
                                    "event_title",
                                    "event_date",
                                    "event_city",
                                    "event_address",
                                    "event_status"
                                ],
                                "properties": {
                                    "event_category": {
                                        "type": "string",
                                        "example": "Music"
                                    },
                                    "event_title": {
                                        "type": "string",
                                        "example": "Concert de Rock"
                                    },
                                    "event_description": {
                                        "type": "string",
                                        "example": "Un concert de rock avec les plus grandes stars."
                                    },
                                    "event_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-09-15T20:00:00Z"
                                    },
                                    "event_image": {
                                        "type": "string",
                                        "example": "https://example.com/images/concert.jpg"
                                    },
                                    "event_city": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "event_address": {
                                        "type": "string",
                                        "example": "Stade de France, 93216 Saint-Denis"
                                    },
                                    "event_status": {
                                        "type": "string",
                                        "example": "upcoming"
                                    },
                                    "event_created_on": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-08-14T12:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Event updated successfully"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Delete an event",
                "description": "Delete a specific event by its ID.",
                "operationId": "DestroyEvent",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID of the event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event deleted successfully"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
...
    },
...
}